feat, design: 쿠폰 등록, 프로젝트 등록, 후원 결제 등 주요 서비스 기능과 공통 컴포넌트, 모달 등 대거 추가·개선
