+++ src/components/projectRegister/StepThreeForm.tsx
@@ -2,7 +2,7 @@
 
 import type React from "react"
 
-import { useState, useEffect } from "react"
+import { useState, useEffect, useRef } from "react"
 import { Plus, Check, X, Trash2 } from "lucide-react"
 import { useProjectFormStore } from "@/stores/useProjectFormStore"
 import type { CompositionItem, SupportOption } from "@/stores/useProjectFormStore"
@@ -10,6 +10,10 @@ import { useApi } from "@/hooks/useApi"
 import { PrimaryButton } from "../common/Button"
 import PlanSelection from "./PlanSelection"
 import AlertModal from "../common/AlertModal"
+import gsap from "gsap"
+import Modal from "../common/Modal"
+import AddAccountModal from "./modals/AddAccountModal"
+import CompositionDialog from "./modals/CompositionDialog"
 
 // 계좌 정보 타입 정의
 interface BankAccount {
@@ -18,268 +22,6 @@ interface BankAccount {
   bankName: string // 은행명
   bankAccount: string // 계좌번호 (키 이름 변경됨)
   isDefault?: boolean // 기본 계좌 여부
-}
-
-// 계좌 추가 모달 컴포넌트
-interface AddAccountModalProps {
-  isOpen: boolean
-  onClose: () => void
-  onSave: (account: Omit<BankAccount, "id" | "isDefault">) => void
-}
-
-const AddAccountModal = ({ isOpen, onClose, onSave }: AddAccountModalProps) => {
-  const [accountHolder, setAccountHolder] = useState("")
-  const [bankName, setBankName] = useState("")
-  const [accountNumber, setAccountNumber] = useState("")
-  const [focusedField, setFocusedField] = useState<string | null>(null)
-  const [errors, setErrors] = useState({
-    accountHolder: false,
-    bankName: false,
-    accountNumber: false,
-  })
-  // 2. AddAccountModal 컴포넌트 내부에 AlertModal 상태 추가
-  const [alertMessage, setAlertMessage] = useState("")
-  const [isAlertOpen, setIsAlertOpen] = useState(false)
-
-  if (!isOpen) return null
-
-  // 계좌번호 입력 처리 함수 추가
-  const handleAccountNumberChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value.replace(/[^0-9]/g, "")
-    setAccountNumber(value)
-  }
-
-  // 폼 제출 처리 함수 추가
-  const handleSubmit = () => {
-    if (validateForm()) {
-      onSave({ accountHolder, bankName, bankAccount: accountNumber })
-      onClose()
-    }
-  }
-
-  // 4. AddAccountModal 컴포넌트 return 문 내부 최상단에 AlertModal 컴포넌트 추가
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      <AlertModal isOpen={isAlertOpen} message={alertMessage} onClose={() => setIsAlertOpen(false)} />
-      {/* 오버레이 */}
-      <div className="fixed inset-0 bg-black/60" onClick={onClose} />
-
-      {/* 모달 */}
-      <div className="relative z-10 w-full max-w-md rounded-3xl bg-white p-6 shadow-lg">
-        <div className="flex items-center justify-between mb-6">
-          <h2 className="text-2xl font-bold">계좌 추가</h2>
-          <button type="button" onClick={onClose} className="text-gray-500 hover:text-gray-700">
-            <X size={24} />
-          </button>
-        </div>
-
-        <div className="space-y-4">
-          {/* 예금주 */}
-          <div>
-            <label
-              htmlFor="accountHolder"
-              className={`text-sm ${focusedField === "accountHolder" ? "text-secondary-color-dark" : "text-main-color"}`}
-            >
-              예금주
-            </label>
-            <input
-              id="accountHolder"
-              type="text"
-              value={accountHolder}
-              onChange={(e) => setAccountHolder(e.target.value)}
-              placeholder="예금주 이름을 입력하세요"
-              className={`w-full p-2 border-b ${
-                focusedField === "accountHolder" ? "border-secondary-color-dark" : "border-gray-300"
-              } ${errors.accountHolder ? "border-red-500" : ""} focus:outline-none text-lg`}
-              onFocus={() => setFocusedField("accountHolder")}
-              onBlur={() => setFocusedField(null)}
-            />
-            {errors.accountHolder && <p className="text-red-500 text-xs mt-1">예금주를 입력해주세요</p>}
-          </div>
-
-          {/* 은행명 */}
-          <div>
-            <label
-              htmlFor="bankName"
-              className={`text-sm ${focusedField === "bankName" ? "text-secondary-color-dark" : "text-main-color"}`}
-            >
-              은행명
-            </label>
-            <input
-              id="bankName"
-              type="text"
-              value={bankName}
-              onChange={(e) => setBankName(e.target.value)}
-              placeholder="은행명을 입력하세요"
-              className={`w-full p-2 border-b ${
-                focusedField === "bankName" ? "border-secondary-color-dark" : "border-gray-300"
-              } ${errors.bankName ? "border-red-500" : ""} focus:outline-none text-lg`}
-              onFocus={() => setFocusedField("bankName")}
-              onBlur={() => setFocusedField(null)}
-            />
-            {errors.bankName && <p className="text-red-500 text-xs mt-1">은행명을 입력해주세요</p>}
-          </div>
-
-          {/* 계좌번호 */}
-          <div>
-            <label
-              htmlFor="accountNumber"
-              className={`text-sm ${focusedField === "accountNumber" ? "text-secondary-color-dark" : "text-main-color"}`}
-            >
-              계좌번호
-            </label>
-            <input
-              id="accountNumber"
-              type="text"
-              value={accountNumber}
-              onChange={handleAccountNumberChange}
-              placeholder="계좌번호를 입력하세요 ('-' 없이 숫자만)"
-              className={`w-full p-2 border-b ${
-                focusedField === "accountNumber" ? "border-secondary-color-dark" : "border-gray-300"
-              } ${errors.accountNumber ? "border-red-500" : ""} focus:outline-none text-lg`}
-              onFocus={() => setFocusedField("accountNumber")}
-              onBlur={() => setFocusedField(null)}
-            />
-            {errors.accountNumber && <p className="text-red-500 text-xs mt-1">계좌번호를 입력해주세요</p>}
-          </div>
-        </div>
-
-        <div className="mt-6 flex justify-end">
-          <button
-            type="button"
-            onClick={handleSubmit}
-            className="bg-main-color hover:bg-secondary-color-dark text-white font-bold py-2 px-6 rounded-full"
-          >
-            추가하기
-          </button>
-        </div>
-      </div>
-    </div>
-  )
-
-  // 3. validateForm 함수 내 alert 호출 부분 수정
-  function validateForm() {
-    const newErrors = {
-      accountHolder: accountHolder.trim() === "",
-      bankName: bankName.trim() === "",
-      accountNumber: accountNumber.trim() === "",
-    }
-
-    setErrors(newErrors)
-    return !Object.values(newErrors).some(Boolean)
-  }
-}
-
-// 구성 다이얼로그 컴포넌트
-interface CompositionDialogProps {
-  isOpen: boolean
-  onClose: () => void
-  composition: CompositionItem[]
-  onSave: (composition: CompositionItem[]) => void
-}
-
-const CompositionDialog = ({ isOpen, onClose, composition, onSave }: CompositionDialogProps) => {
-  const [items, setItems] = useState<CompositionItem[]>(composition || [])
-  const [focusedField, setFocusedField] = useState<string | null>(null)
-
-  if (!isOpen) return null
-
-  const addItem = () => {
-    const newId = items.length > 0 ? Math.max(...items.map((item) => item.id)) + 1 : 1
-    setItems([...items, { id: newId, content: "" }])
-  }
-
-  const removeItem = (id: number) => {
-    setItems(items.filter((item) => item.id !== id))
-  }
-
-  const updateItem = (id: number, content: string) => {
-    setItems(items.map((item) => (item.id === id ? { ...item, content } : item)))
-  }
-
-  const handleSave = () => {
-    onSave(items)
-    onClose()
-  }
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center">
-      {/* 오버레이 */}
-      <div className="fixed inset-0 bg-black/60" onClick={onClose} />
-
-      {/* 다이얼로그 */}
-      <div className="relative z-10 w-full max-w-2xl rounded-3xl bg-white p-6 shadow-lg">
-        <div className="flex items-center justify-between mb-6">
-          <h2 className="text-2xl font-bold">구성 항목 설정</h2>
-          <button type="button" onClick={onClose} className="text-gray-500 hover:text-gray-700">
-            <X size={24} />
-          </button>
-        </div>
-
-        <div className="space-y-4 max-h-[60vh] overflow-y-auto">
-          {items.map((item) => (
-            <div key={item.id} className="flex items-end gap-4">
-              <div className="flex-1">
-                <label
-                  htmlFor={`item-content-${item.id}`}
-                  className={`text-sm ${focusedField === `item-content-${item.id}` ? "text-secondary-color-dark" : "text-main-color"}`}
-                >
-                  구성 세부 내용 (100자 이하)
-                </label>
-                <input
-                  id={`item-content-${item.id}`}
-                  type="text"
-                  value={item.content}
-                  onChange={(e) => {
-                    // 100자 제한
-                    if (e.target.value.length <= 100) {
-                      updateItem(item.id, e.target.value)
-                    }
-                  }}
-                  placeholder="구성 세부 내용을 입력하세요"
-                  className={`w-full p-2 border-b ${
-                    focusedField === `item-content-${item.id}` ? "border-secondary-color-dark" : "border-gray-300"
-                  } focus:outline-none text-lg`}
-                  onFocus={() => setFocusedField(`item-content-${item.id}`)}
-                  onBlur={() => setFocusedField(null)}
-                />
-              </div>
-
-              <button
-                type="button"
-                onClick={() => removeItem(item.id)}
-                className="p-2 text-gray-400 hover:text-red-500 transition-colors mb-2"
-                aria-label="항목 삭제"
-              >
-                <Trash2 className="h-5 w-5" />
-              </button>
-            </div>
-          ))}
-
-          {/* 항목 추가 버튼 */}
-          <div
-            className="border border-dashed border-gray-300 rounded-xl p-3 flex items-center justify-center cursor-pointer hover:bg-gray-50 h-14"
-            onClick={addItem}
-          >
-            <div className="flex items-center text-gray-500">
-              <Plus className="w-5 h-5 mr-2" />
-              <span className="text-sm font-medium">구성 항목 추가</span>
-            </div>
-          </div>
-        </div>
-
-        <div className="mt-6 flex justify-end">
-          <PrimaryButton
-            type="button"
-            onClick={handleSave}
-            className="bg-main-color text-white font-bold py-2 px-6 rounded-full"
-          >
-            저장
-          </PrimaryButton>
-        </div>
-      </div>
-    </div>
-  )
 }
 
 interface StepThreeFormProps {
@@ -517,88 +259,100 @@ export default function StepThreeForm({ initialData, isEditMode = false }: StepT
               key={option.id}
               className="border border-gray-300 rounded-3xl p-4 h-80 flex flex-col justify-between relative"
             >
-              {/* 삭제 버튼 추가 */}
-              <button
-                type="button"
-                onClick={() => removeSupportOption(option.id)}
-                className="absolute right-4 top-4 p-1.5 text-gray-400 hover:text-red-500 transition-colors rounded-full hover:bg-gray-100"
-                aria-label="옵션 삭제"
-              >
-                <Trash2 className="h-5 w-5" />
-              </button>
+              {/* 삭제 버튼은 수정 모드에서 비활성화 */}
+              {!isEditMode && (
+                <button
+                  type="button"
+                  onClick={() => removeSupportOption(option.id)}
+                  className="absolute right-4 top-4 p-1.5 text-gray-400 hover:text-red-500 transition-colors rounded-full hover:bg-gray-100"
+                  aria-label="옵션 삭제"
+                >
+                  <Trash2 className="h-5 w-5" />
+                </button>
+              )}
 
               <div className="flex flex-col space-y-4">
                 {/* 옵션 이름 */}
                 <div>
                   <label
                     htmlFor={`option-name-${option.id}`}
-                    className={`text-sm ${focusedField === `option-name-${option.id}` ? "text-secondary-color-dark" : "text-main-color"}`}
+                    className={`text-sm ${isEditMode ? "text-gray-400" : (focusedField === `option-name-${option.id}` ? "text-secondary-color-dark" : "text-main-color")}`}
                   >
                     상품 이름 (25자 이하)
                   </label>
-                  <input
-                    id={`option-name-${option.id}`}
-                    type="text"
-                    value={option.name}
-                    onChange={(e) => {
-                      // 25자 제한
-                      if (e.target.value.length <= 25) {
-                        updateSupportOption(option.id, "name", e.target.value)
-                      }
-                    }}
-                    placeholder="상품 이름을 지어주세요."
-                    className={`w-full p-1 border-b ${
-                      focusedField === `option-name-${option.id}` ? "border-secondary-color-dark" : "border-gray-300"
-                    } focus:outline-none text-base`}
-                    onFocus={() => setFocusedField(`option-name-${option.id}`)}
-                    onBlur={() => setFocusedField(null)}
-                  />
+                  {isEditMode ? (
+                    <div className="w-full p-1 border-b border-gray-200 text-base text-gray-400 cursor-not-allowed">{option.name}</div>
+                  ) : (
+                    <input
+                      id={`option-name-${option.id}`}
+                      type="text"
+                      value={option.name}
+                      onChange={(e) => {
+                        if (e.target.value.length <= 25) {
+                          updateSupportOption(option.id, "name", e.target.value)
+                        }
+                      }}
+                      placeholder="상품 이름을 지어주세요."
+                      className={`w-full p-1 border-b ${focusedField === `option-name-${option.id}` ? "border-secondary-color-dark" : "border-gray-300"} focus:outline-none text-base`}
+                      onFocus={() => setFocusedField(`option-name-${option.id}`)}
+                      onBlur={() => setFocusedField(null)}
+                    />
+                  )}
                 </div>
 
                 {/* 옵션 설명 */}
                 <div>
                   <label
                     htmlFor={`option-desc-${option.id}`}
-                    className={`text-sm ${focusedField === `option-desc-${option.id}` ? "text-secondary-color-dark" : "text-main-color"}`}
+                    className={`text-sm ${isEditMode ? "text-gray-400" : (focusedField === `option-desc-${option.id}` ? "text-secondary-color-dark" : "text-main-color")}`}
                   >
                     상품 설명 (50자 이하)
                   </label>
-                  <input
-                    id={`option-desc-${option.id}`}
-                    type="text"
-                    value={option.description}
-                    onChange={(e) => {
-                      // 50자 제한
-                      if (e.target.value.length <= 50) {
-                        updateSupportOption(option.id, "description", e.target.value)
-                      }
-                    }}
-                    placeholder="해당 옵션을 자세히 설명해 주세요."
-                    className={`w-full p-1 border-b ${
-                      focusedField === `option-desc-${option.id}` ? "border-secondary-color-dark" : "border-gray-300"
-                    } focus:outline-none text-base`}
-                    onFocus={() => setFocusedField(`option-desc-${option.id}`)}
-                    onBlur={() => setFocusedField(null)}
-                  />
+                  {isEditMode ? (
+                    <div className="w-full p-1 border-b border-gray-200 text-base text-gray-400 cursor-not-allowed">{option.description}</div>
+                  ) : (
+                    <input
+                      id={`option-desc-${option.id}`}
+                      type="text"
+                      value={option.description}
+                      onChange={(e) => {
+                        if (e.target.value.length <= 50) {
+                          updateSupportOption(option.id, "description", e.target.value)
+                        }
+                      }}
+                      placeholder="해당 옵션을 자세히 설명해 주세요."
+                      className={`w-full p-1 border-b ${focusedField === `option-desc-${option.id}` ? "border-secondary-color-dark" : "border-gray-300"} focus:outline-none text-base`}
+                      onFocus={() => setFocusedField(`option-desc-${option.id}`)}
+                      onBlur={() => setFocusedField(null)}
+                    />
+                  )}
                 </div>
 
                 {/* 구성 */}
                 <div>
-                  <label className="text-sm text-main-color">구성</label>
-                  <div className="mt-0.5">
-                    <button
-                      type="button"
-                      onClick={() => openCompositionDialog(option.id)}
-                      className="flex items-center text-main-color hover:text-secondary-color-dark"
-                    >
-                      <span>
-                        구성 항목 추가/편집...
-                        {option.composition && option.composition.length > 0 && (
-                          <span className="ml-1 text-gray-500">({option.composition.length}개 추가됨)</span>
-                        )}
-                      </span>
-                    </button>
-                  </div>
+                  <label className={`text-sm ${isEditMode ? "text-gray-400" : "text-main-color"}`}>구성</label>
+                  {isEditMode ? (
+                    <div className="mt-0.5 min-h-[24px] text-base text-gray-400 border-b border-gray-200 p-1 cursor-not-allowed">
+                      {option.composition && option.composition.length > 0
+                        ? option.composition.map((item) => item.content).join(", ")
+                        : "구성 없음"}
+                    </div>
+                  ) : (
+                    <div className="mt-0.5">
+                      <button
+                        type="button"
+                        onClick={() => openCompositionDialog(option.id)}
+                        className="flex items-center text-main-color hover:text-secondary-color-dark"
+                      >
+                        <span>
+                          구성 항목 추가/편집...
+                          {option.composition && option.composition.length > 0 && (
+                            <span className="ml-1 text-gray-500">({option.composition.length}개 추가됨)</span>
+                          )}
+                        </span>
+                      </button>
+                    </div>
+                  )}
                 </div>
               </div>
 
@@ -608,26 +362,33 @@ export default function StepThreeForm({ initialData, isEditMode = false }: StepT
                 <div className="flex-1">
                   <label
                     htmlFor={`option-price-${option.id}`}
-                    className={`text-sm ${focusedField === `option-price-${option.id}` ? "text-secondary-color-dark" : "text-main-color"}`}
+                    className={`text-sm ${isEditMode ? "text-gray-400" : (focusedField === `option-price-${option.id}` ? "text-secondary-color-dark" : "text-main-color")}`}
                   >
                     가격 (10억 원 이하)
                   </label>
-                  <div className="flex items-center border-b border-gray-300 focus-within:border-secondary-color-dark">
-                    <input
-                      id={`option-price-${option.id}`}
-                      type="text"
-                      value={formatNumber(option.price)}
-                      onChange={(e) => handlePriceChange(option.id, e.target.value)}
-                      placeholder="0"
-                      className="w-full p-1 focus:outline-none text-base text-right"
-                      onFocus={() => setFocusedField(`option-price-${option.id}`)}
-                      onBlur={() => setFocusedField(null)}
-                    />
-                    <span className="text-gray-500 ml-1">원</span>
-                  </div>
+                  {isEditMode ? (
+                    <div className="flex items-center border-b border-gray-200 text-base text-gray-400 cursor-not-allowed p-1">
+                      <span className="flex-1 text-right">{formatNumber(option.price)}</span>
+                      <span className="text-gray-400 ml-1">원</span>
+                    </div>
+                  ) : (
+                    <div className="flex items-center border-b border-gray-300 focus-within:border-secondary-color-dark">
+                      <input
+                        id={`option-price-${option.id}`}
+                        type="text"
+                        value={formatNumber(option.price)}
+                        onChange={(e) => handlePriceChange(option.id, e.target.value)}
+                        placeholder="0"
+                        className="w-full p-1 focus:outline-none text-base text-right"
+                        onFocus={() => setFocusedField(`option-price-${option.id}`)}
+                        onBlur={() => setFocusedField(null)}
+                      />
+                      <span className="text-gray-500 ml-1">원</span>
+                    </div>
+                  )}
                 </div>
 
-                {/* 재고 */}
+                {/* 재고(수량) */}
                 <div className="flex-1">
                   <label
                     htmlFor={`option-stock-${option.id}`}
@@ -781,18 +542,16 @@ export default function StepThreeForm({ initialData, isEditMode = false }: StepT
       />
 
       {/* 구성 다이얼로그 */}
-      {compositionDialogOpen && (
-        <CompositionDialog
-          isOpen={compositionDialogOpen}
-          onClose={() => setCompositionDialogOpen(false)}
-          composition={
-            currentOptionId !== null
-              ? supportOptions.find((option) => option.id === currentOptionId)?.composition || []
-              : []
-          }
-          onSave={saveComposition}
-        />
-      )}
+      <CompositionDialog
+        isOpen={compositionDialogOpen}
+        onClose={() => setCompositionDialogOpen(false)}
+        composition={
+          currentOptionId !== null
+            ? supportOptions.find((option) => option.id === currentOptionId)?.composition || []
+            : []
+        }
+        onSave={saveComposition}
+      />
     </div>
   )
